Fix the tracker structure to store trackers linked to the TrackerCore pointer. Implement start/stop timer per tracker.

+++ app/src/main/java/com/newrelic/videotest/Exo2Activity.java
@@ -44,7 +44,7 @@ public class Exo2Activity extends AppCompatActivity {
 
     private SimpleExoPlayer player;
     private CastPlayer castPlayer;
-    private Integer trackerID;
+    private Long trackerID;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
